+++ src/Controller/Admin/CountryController.php
@@ -11,6 +11,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\String\Slugger\SluggerInterface;
 
 class CountryController extends AbstractController
 {
@@ -18,7 +19,8 @@ class CountryController extends AbstractController
     public function createCountry(Request                 $request,
                                   EntityManagerInterface  $entityManager,
                                   ParameterBagInterface   $parameterBag,
-                                  UniqueFilenameGenerator $uniqueFilenameGenerator)
+                                  UniqueFilenameGenerator $uniqueFilenameGenerator,
+                                  SluggerInterface        $slugger)
     {
         $country = new Country();
 
@@ -29,6 +31,9 @@ class CountryController extends AbstractController
         $formCountry->handleRequest($request);
 
         if ($formCountry->isSubmitted() && $formCountry->isValid()) {
+
+            $country->setSlugFromName($slugger);
+
             $countryFlag = $formCountry->get('flag')->getData();
             $countryBanner = $formCountry->get('banner')->getData();
 
@@ -86,26 +91,27 @@ class CountryController extends AbstractController
         ]);
     }
 
-    #[Route('/admin/country/show/{id}', name: 'country_show', methods: ['GET'])]
-    public function showCountry(int $id, CountryRepository $countryRepository)
+    #[Route('/admin/country/show/{slug}', name: 'country_show', methods: ['GET'])]
+    public function showCountry(string $slug, CountryRepository $countryRepository)
     {
-        $country = $countryRepository->find($id);
+        $country = $countryRepository->findOneBy(['slug' => $slug]);
 
         return $this->render('admin/country/show.html.twig', [
             'country' => $country
         ]);
     }
 
-    #[Route('/admin/country/update/{id}', name: 'country_update', methods: ['GET', 'POST'])]
-    public function updateCountry(int                     $id,
+    #[Route('/admin/country/update/{slug}', name: 'country_update', methods: ['GET', 'POST'])]
+    public function updateCountry(string                     $slug,
                                   Request                 $request,
                                   EntityManagerInterface  $entityManager,
                                   CountryRepository       $countryRepository,
                                   UniqueFilenameGenerator $uniqueFilenameGenerator,
                                   ParameterBagInterface   $parameterBag,
+                                  SluggerInterface        $slugger
     )
     {
-        $country = $countryRepository->find($id);
+        $country = $countryRepository->findOneBy(['slug' => $slug]);
 
         $formCountry = $this->createForm(CountryType::class, $country, [
             'is_require' => true
@@ -114,6 +120,9 @@ class CountryController extends AbstractController
         $formCountry->handleRequest($request);
 
         if ($formCountry->isSubmitted() && $formCountry->isValid()) {
+
+            $country->setSlugFromName($slugger);
+
             $countryFlag = $formCountry->get('flag')->getData();
             $countryBanner = $formCountry->get('banner')->getData();
 
@@ -181,10 +190,14 @@ class CountryController extends AbstractController
 
     }
 
-    #[Route('/admin/country/delete/{id}', name: 'country_delete', methods: ['GET'])]
-    public function deleteCountry(int $id, EntityManagerInterface $entityManager, CountryRepository $countryRepository, ParameterBagInterface $parameterBag)
+    #[Route('/admin/country/delete/{slug}', name: 'country_delete', methods: ['GET', 'POST'])]
+    public function deleteCountry(string $slug,
+                                  EntityManagerInterface $entityManager,
+                                  CountryRepository $countryRepository,
+                                  SluggerInterface $slugger,
+                                  ParameterBagInterface $parameterBag)
     {
-        $country = $countryRepository->find($id);
+        $country = $countryRepository->findOneBy(['slug' => $slug]);
 
         $projectDir = $parameterBag->get('kernel.project_dir');