+++ src/Entity/Continent.php
@@ -6,8 +6,11 @@ use App\Repository\ContinentRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
+use Symfony\Component\String\Slugger\SluggerInterface;
 
 #[ORM\Entity(repositoryClass: ContinentRepository::class)]
+#[UniqueEntity(fields: ['slug'])]
 class Continent
 {
     #[ORM\Id]
@@ -20,6 +23,9 @@ class Continent
 
     #[ORM\OneToMany(mappedBy: 'continent', targetEntity: Country::class, orphanRemoval: true)]
     private Collection $countries;
+
+    #[ORM\Column(length: 255, unique: true, nullable: false)]
+    private ?string $slug = null;
 
     public function __construct()
     {
@@ -71,5 +77,22 @@ class Continent
         }
 
         return $this;
+    }
+
+    public function getSlug(): ?string
+    {
+        return $this->slug;
+    }
+
+    public function setSlug(string $slug): static
+    {
+        $this->slug = $slug;
+
+        return $this;
+    }
+
+    public function setSlugFromName(SluggerInterface $slugger): void
+    {
+        $this->slug = strtolower($slugger->slug($this->name));
     }
 }