+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.17](https://github.com/nathkeuss/GEONERD/compare/v0.1.16...v0.1.17) (2025-02-20)
+
+
+### Features
+
+* **country:** implement slug functionality for country entity ([498bb53](https://github.com/nathkeuss/GEONERD/commit/498bb537418a56dc195fccbbbd607e62c79ac93d))
+
+### [0.1.16](https://github.com/nathkeuss/GEONERD/compare/v0.1.15...v0.1.16) (2025-02-20)
+
+
+### Features
+
+* **continent:** implement slug functionality for continent entity ([46e7d0c](https://github.com/nathkeuss/GEONERD/commit/46e7d0c6e78dcbd64040ad5763e7f97eb781b619))
+
 ### [0.1.15](https://github.com/nathkeuss/GEONERD/compare/v0.1.14...v0.1.15) (2025-02-19)